requirements.txt
ArchivesSnake==0.2.2
attrs==17.4.0
boltons==18.0.0
certifi==2018.4.16
chardet==3.0.4
click==6.7
clinner==1.9.4
colorama==0.3.9
colorlog==3.1.4
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.1
Django==2.0
django-common-helpers==0.9.1
django-cron==0.5.0
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
drf-yasg==1.7.4
flex==6.13.2
future==0.16.0
gitdb2==2.0.3
GitPython==2.1.9
health-check==3.4.1
idna==2.6
inflection==0.3.1
itypes==1.1.0
Jinja2==2.10
jsonpointer==1.14
jsonschema==2.6.0
Markdown==2.6.11
MarkupSafe==1.0
multidict==4.2.0
mysqlclient==1.3.12
openapi-codec==1.3.2
PyJWT==1.6.1
pytz==2018.4
PyYAML==3.12
requests==2.18.4
rfc3987==1.3.7
ruamel.yaml==0.15.37
six==1.11.0
smmap2==2.0.3
strict-rfc3339==0.7
structlog==18.1.0
swagger-spec-validator==2.1.0
uritemplate==3.0.0
urllib3==1.22
uuid==1.30
validate-email==1.3
vcrpy==1.11.1
wrapt==1.10.11
yarl==1.1.1
